Switch to decimal and prevent double flags

+++ MazeMakerUtilities/Parameters.cs
@@ -4,18 +4,18 @@ namespace MazeMakerUtilities
 {
     public class Parameters
     {
-        public float XHorizontalStart { get; set; }
-        public float XHorizontalOffset { get; set; }
-        public float ZHorizontalStart { get; set; }
-        public float ZHorizontalOffset { get; set; }
-        public float YHorizontalStart { get; set; }
+        public decimal XHorizontalStart { get; set; }
+        public decimal XHorizontalOffset { get; set; }
+        public decimal ZHorizontalStart { get; set; }
+        public decimal ZHorizontalOffset { get; set; }
+        public decimal YHorizontalStart { get; set; }
 
 
-        public float XVerticalStart { get; set; }
-        public float XVerticalOffset { get; set; }
-        public float ZVerticalStart { get; set; }
-        public float ZVerticalOffset { get; set; }
-        public float YVerticalStart { get; set; }
+        public decimal XVerticalStart { get; set; }
+        public decimal XVerticalOffset { get; set; }
+        public decimal ZVerticalStart { get; set; }
+        public decimal ZVerticalOffset { get; set; }
+        public decimal YVerticalStart { get; set; }
 
         public int gridColumns { get; set; }
         public int gridRows { get; set; }
@@ -49,17 +49,17 @@ namespace MazeMakerUtilities
             if (parameters.horizontalWide && parameters.verticalWide)
             {
                 parameters.gridColumns = 9;
-                parameters.XVerticalStart = 12.5F;
-                parameters.XVerticalOffset = 3.2F;
-                parameters.ZVerticalStart = 14.2F;
-                parameters.ZVerticalOffset = 3.0F;
+                parameters.XVerticalStart = 12.5M;
+                parameters.XVerticalOffset = 3.2M;
+                parameters.ZVerticalStart = 14.2M;
+                parameters.ZVerticalOffset = 3.0M;
                 parameters.YVerticalStart = 0;
 
                 parameters.gridRows = 9;
-                parameters.XHorizontalStart = 14.1F;
-                parameters.XHorizontalOffset = 3.2F;
-                parameters.ZHorizontalStart = 12.8F;
-                parameters.ZHorizontalOffset = 3F;
+                parameters.XHorizontalStart = 14.1M;
+                parameters.XHorizontalOffset = 3.2M;
+                parameters.ZHorizontalStart = 12.8M;
+                parameters.ZHorizontalOffset = 3M;
                 parameters.YHorizontalStart = 0;
 
                 parameters.entranceRow = 8;
@@ -72,17 +72,17 @@ namespace MazeMakerUtilities
             else if (!parameters.horizontalWide && !parameters.verticalWide)
             {
                 parameters.gridColumns = 28;
-                parameters.XVerticalStart = 13.5F;
-                parameters.XVerticalOffset = 1.2F;
-                parameters.ZVerticalStart = 14.7F;
-                parameters.ZVerticalOffset = 1.0F;
+                parameters.XVerticalStart = 13.5M;
+                parameters.XVerticalOffset = 1.2M;
+                parameters.ZVerticalStart = 14.7M;
+                parameters.ZVerticalOffset = 1.0M;
                 parameters.YVerticalStart = 0;
 
                 parameters.gridRows = 25;
-                parameters.XHorizontalStart = 14.1F;
-                parameters.XHorizontalOffset = 1.2F;
-                parameters.ZHorizontalStart = 14.3F;
-                parameters.ZHorizontalOffset = 1F;
+                parameters.XHorizontalStart = 14.1M;
+                parameters.XHorizontalOffset = 1.2M;
+                parameters.ZHorizontalStart = 14.3M;
+                parameters.ZHorizontalOffset = 1M;
                 parameters.YHorizontalStart = 0;
 
                 parameters.entranceRow = 24;
@@ -95,17 +95,17 @@ namespace MazeMakerUtilities
             else if (parameters.horizontalWide &&  !parameters.verticalWide)
             {
                 parameters.gridColumns = 9;                
-                parameters.XVerticalStart = 13.5F;
-                parameters.XVerticalOffset = 1.2F;
-                parameters.ZVerticalStart = 14.0F;
-                parameters.ZVerticalOffset = 3.0F;
+                parameters.XVerticalStart = 13.5M;
+                parameters.XVerticalOffset = 1.2M;
+                parameters.ZVerticalStart = 14.0M;
+                parameters.ZVerticalOffset = 3.0M;
                 parameters.YVerticalStart = 0;
 
                 parameters.gridRows = 25;
-                parameters.XHorizontalStart = 14.1F;
-                parameters.XHorizontalOffset = 1.2F;
-                parameters.ZHorizontalStart = 12.6F;
-                parameters.ZHorizontalOffset = 3F;
+                parameters.XHorizontalStart = 14.1M;
+                parameters.XHorizontalOffset = 1.2M;
+                parameters.ZHorizontalStart = 12.6M;
+                parameters.ZHorizontalOffset = 3M;
                 parameters.YHorizontalStart = 0;
 
                 parameters.entranceRow = 24;
@@ -119,17 +119,17 @@ namespace MazeMakerUtilities
             else if (!parameters.horizontalWide && parameters.verticalWide)
             {
                 parameters.gridColumns = 23;
-                parameters.XVerticalStart = 12.7F;
-                parameters.XVerticalOffset = 3.0F;
-                parameters.ZVerticalStart = 14.7F;
-                parameters.ZVerticalOffset = 1.2F;
+                parameters.XVerticalStart = 12.7M;
+                parameters.XVerticalOffset = 3.0M;
+                parameters.ZVerticalStart = 14.7M;
+                parameters.ZVerticalOffset = 1.2M;
                 parameters.YVerticalStart = 0;
 
                 parameters.gridRows = 10;
-                parameters.XHorizontalStart = 14.3F;
-                parameters.XHorizontalOffset = 3.0F;
-                parameters.ZHorizontalStart = 14.1F;
-                parameters.ZHorizontalOffset = 1.2F;
+                parameters.XHorizontalStart = 14.3M;
+                parameters.XHorizontalOffset = 3.0M;
+                parameters.ZHorizontalStart = 14.1M;
+                parameters.ZHorizontalOffset = 1.2M;
                 parameters.YHorizontalStart = 0;
 
                 parameters.entranceRow = 9;
